feat: adicionados DTO's para lidar interagir com as requisições - refatorado o atributo nivel da entidade User

+++ src/main/java/com/felipe_dias/backend_java_spring_test/Service/impl/UserServiceImp.java
@@ -2,6 +2,7 @@ package com.felipe_dias.backend_java_spring_test.Service.impl;
 
 import com.felipe_dias.backend_java_spring_test.Controller.Exceptions.ResourceNotFoundException;
 import com.felipe_dias.backend_java_spring_test.Service.UserService;
+import com.felipe_dias.backend_java_spring_test.model.Task;
 import com.felipe_dias.backend_java_spring_test.model.User;
 import com.felipe_dias.backend_java_spring_test.model.dto.UserDTO;
 import com.felipe_dias.backend_java_spring_test.repository.UserRepository;
@@ -9,6 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class UserServiceImp implements UserService {
@@ -17,8 +19,13 @@ public class UserServiceImp implements UserService {
     private UserRepository userRepository;
 
     @Override
-    public List<User> getAllUsers() {
-        return userRepository.findAll();
+    public List<UserDTO> getAllUsers() {
+        List<User> usersList = userRepository.findAll();
+
+        List<UserDTO> dtoList = usersList.stream()
+                                        .map(user -> new UserDTO(user))
+                                        .collect(Collectors.toList());
+        return dtoList;
     }
 
     @Override
@@ -28,9 +35,12 @@ public class UserServiceImp implements UserService {
             throw new IllegalArgumentException("CAMPO USERNAME OBRIGATORIO");
         }
 
-        if(user.getNivel().isEmpty() || user.getNivel() == null){
+        if(user.getNivel().getLabel().isEmpty() || user.getNivel() == null){
             throw new IllegalArgumentException("CAMPO NIVEL OBRIGATORIO");
         }
+        //TODO: LIDAR COM O NIVEL DO USUARIO
+
+
         userRepository.save(user);
     }
 
@@ -47,14 +57,19 @@ public class UserServiceImp implements UserService {
             foundUser.setUsername(user.getUsername());
         }
 
-
-
+        userRepository.save(foundUser);
 
     }
 
     @Override
     public void deleteUser(Long id) {
 
+        if(!userRepository.existsById(id)){
+            throw new ResourceNotFoundException(id);
+        }
+
         userRepository.deleteById(id);
     }
+
+
 }
